code

+++ automation/src/main/java/ocean/test/condition/pricing/OCEAN_Pricing_TC_23_24.java
@@ -1,13 +1,12 @@
 package ocean.test.condition.pricing;
 
-import static org.testng.Assert.assertNotEquals;
+import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
 import java.util.HashMap;
 
 import org.testng.annotations.Test;
 
-import ocean.modules.dataprovider.PricingDataProvider;
 import ocean.modules.pages.PricingModulePages;
 
 /**
@@ -17,6 +16,8 @@ import ocean.modules.pages.PricingModulePages;
  * Description: Validate classification list for cloning/exporting and delete.
  * 
  * @author Shalu Chauhan
+ * 
+ * @reviewer : Shalu Chauhan
  */
 
 public class OCEAN_Pricing_TC_23_24 extends PricingModulePages {
@@ -24,49 +25,51 @@ public class OCEAN_Pricing_TC_23_24 extends PricingModulePages {
 	 * Validate classification list for cloning/exporting and delete.this function
 	 * cover the condition of this test case like 1. Cloning ClassificationList
 	 */
-	@Test(priority = 1, groups = "regression", dataProvider = "fetchDataForTC24", dataProviderClass = PricingDataProvider.class, description = "Validate classification list for cloning.")
-	public void validateClassificationListForCloneExport(String[] inputArray) throws Exception {
+	@Test(priority = 5, groups = { "regression", "extendSmoke",
+			"fullSuite" }, description = "Validate classification list for cloning.")
+	public void validateClassificationListForCloneExport() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
-		HashMap<String, String> map = new HashMap<String, String>();
-		map.put("PriceSheetName", inputArray[0]);
-		selectCloneClassfication(map);
-		validateClonePriceSheet(map);
+		selectCloneClassfication("ClonePriceSheet");
+		validateClonePriceSheet("ClonePriceSheet");
 		String validationMessage = getValueofClonePriceSheet();
-		assertTrue(validationMessage.equalsIgnoreCase(map.get("PriceSheetName")), "Not matched");
+		assertTrue(validationMessage.equalsIgnoreCase("ClonePriceSheet"));
 	}
 
 	/**
 	 * Validate classification list for cloning/exporting and delete.this function
 	 * cover the condition of this test case like 2. Export ClassificationList
 	 */
-	// @Test(priority = 2, groups = "regression", description = "Validate
-	// classification list for exporting.")
+	@Test(priority = 5, groups = { "regression", "extendSmoke",
+			"fullSuite" }, description = "Validate classification list for exporting.")
 	public void validateClassificationListForExport() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
 		exportClassificationPriceSheet();
-		String downloadPath = "C:\\Users\\shalu.chauhan\\Desktop";
+		String desktopPath = System.getProperty("user.home") + "\\Desktop";
 		String fileName = getTextOfElement("getExportClassificationListName");
-		isFileDownloaded(downloadPath, fileName);
-		String deletepath = "C:\\Users\\shalu.chauhan\\Desktop\\Absolute Lease - ACL_Master_Export.xlxs";
-		deleteExportFile(deletepath);
+		boolean deleted = isFileDownloaded(desktopPath, fileName);
+		if (deleted == true) {
+			String deletepath = desktopPath + "\\" + fileName;
+			deleteExportFile(deletepath);
+			boolean deleted1 = isFileDownloaded(desktopPath, fileName);
+			assertEquals(deleted1, false);
+		} else {
+			throw new Exception("File not deleted");
+		}
 	}
 
 	/**
 	 * Validate classification list for cloning/exporting and delete.this function
 	 * cover the condition of this test case like 3. Delete ClassificationList
 	 */
-	// @Test(priority = 3, groups = "regression", dataProvider = "fetchDataForTC24",
-	// dataProviderClass = PricingDataProvider.class, description = "Validate
-	// classification list for delete.")
-	public void validateClassificationListForDelete(String[] inputArray) throws Exception {
-		goToPricingTab();
-		goToClassficationList();
-		HashMap<String, String> map = new HashMap<String, String>();
-		map.put("PriceSheetName", inputArray[0]);
-		deleteClassificationPriceSheet(map.get("PriceSheetName"));
-		validateDeleteClassificationPriceSheet("listOfClassificationList");
+	@Test(dependsOnMethods = { "validateClassificationListForCloneExport" }, priority = 5, groups = { "regression",
+			"extendSmoke", "fullSuite" }, description = "Validate classification list for delete.")
+	public void validateClassificationListForDelete() throws Exception {
+		deleteClassificationPriceSheet("ClonePriceSheet");
+		boolean flag = validateDeleteClassificationPriceSheet("listOfClassificationList");
+		type("typeForFilterClassificationList", " ");
+		assertEquals(flag, true);
 	}
 
 	/**
@@ -75,23 +78,34 @@ public class OCEAN_Pricing_TC_23_24 extends PricingModulePages {
 	 * classification list when it is increased by one on a vehicle for turbo
 	 * option.
 	 */
-	// @Test(priority = 4, groups = "regression", dataProvider = "fetchDataForTC24",
-	// dataProviderClass = PricingDataProvider.class, description = "Validate usage
-	// of classification list when it is increased by one on a vehicle for turbo
-	// option.")
-	public void validateClassificationListForTurbo(String[] inputArray) throws Exception {
+	@Test(priority = 5, groups = "fullSuite", description = "Validate usage of classification list when it is increased by one on a vehicle for turbo option.")
+	public void validateClassificationListForTurbo() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
-		HashMap<String, String> map = new HashMap<String, String>();
-		map.put("Turbo", inputArray[2]);
-		map.put("ClassificationListForTurbo", inputArray[1]);
-		selectPricesheetForTurbo(map);
-		HashMap<String, String> actualTurboValue = getClassficationListForTurboBeforAndAfterEditing(inputArray[1], 0);
-		editClassificationListForTurbo(map);
-		HashMap<String, String> expectedTurboValue = getClassficationListForTurboBeforAndAfterEditing(inputArray[1], 1);
-		uncheckedEditTurboCheckbox();
-		waitForSomeTime(3);
-		assertNotEquals(actualTurboValue, expectedTurboValue);
+		HashMap<String, String> actualTurboValue = getClassificationInfomation("", "0", "0");
+		selectPricesheetForTurbo(actualTurboValue.get("CLASS_GROUP_NAME"));
+		editClassificationListForTurboAndDiesel("Turbo", "Y");
+		String checkboxYes = getAttributeValue("clickCheckBoxOfTurbo", "Toggle.ToggleState");
+		if (checkboxYes.equalsIgnoreCase("1")) {
+			HashMap<String, String> expectedTurboValue = getClassificationInfomation(
+					actualTurboValue.get("CLASS_GROUP_NAME"), "1", "0");
+			if (expectedTurboValue.size() > 1) {
+				rightClick("clickOnCheckbox");
+				editClassificationListForTurboAndDiesel("Turbo", "N");
+				String checkboxNo = getAttributeValue("clickCheckBoxOfTurbo", "Toggle.ToggleState");
+				HashMap<String, String> expectedTurboValueFinal = getClassificationInfomation(
+						actualTurboValue.get("CLASS_GROUP_NAME"), "0", "0");
+				if (expectedTurboValueFinal.size() > 1)
+					assertEquals(checkboxNo, "0");
+				else
+					assertEquals(false, true);
+
+			} else {
+				throw new Exception("Case failed");
+			}
+		} else {
+			throw new Exception("checkbox is not clicked");
+		}
 	}
 
 	/**
@@ -100,24 +114,33 @@ public class OCEAN_Pricing_TC_23_24 extends PricingModulePages {
 	 * classification list when it is increased by one on a vehicle for turbo
 	 * option.
 	 */
-	// @Test(priority = 5, groups = "regression", dataProvider = "fetchDataForTC24",
-	// dataProviderClass = PricingDataProvider.class, description = "Validate usage
-	// of classification list when it is increased by one on a vehicle for turbo
-	// option.")
-	public void validateClassificationListForDiesel(String[] inputArray) throws Exception {
+	@Test(priority = 5, groups = "fullSuite", description = "Validate usage of classification list when it is increased by one on a vehicle for turbo option.")
+	public void validateClassificationListForDiesel() throws Exception {
 		goToPricingTab();
 		goToClassficationList();
-		HashMap<String, String> map1 = new HashMap<String, String>();
-		map1.put("Diesel", inputArray[3]);
-		map1.put("ClassificationListForDiesel", inputArray[4]);
-		selectPricesheetForDiesel(map1);
-		HashMap<String, String> actualDieselValue = getClassficationListForDieselBeforeAndAfterEditing(inputArray[4],
-				0);
-		editClassificationListForDiesel(map1);
-		HashMap<String, String> expectedDieselValue = getClassficationListForDieselBeforeAndAfterEditing(inputArray[4],
-				1);
-		uncheckedEditDieselCheckbox();
-		waitForSomeTime(3);
-		assertNotEquals(actualDieselValue, expectedDieselValue);
+		HashMap<String, String> actualTurboValue = getClassificationInfomation("", "0", "0");
+		selectPricesheetForTurbo(actualTurboValue.get("CLASS_GROUP_NAME"));
+		editClassificationListForTurboAndDiesel("diesel", "Y");
+		String checkboxYes = getAttributeValue("clickCheckBoxOfDiesel", "Toggle.ToggleState");
+		if (checkboxYes.equalsIgnoreCase("1")) {
+			HashMap<String, String> expectedTurboValue = getClassificationInfomation(
+					actualTurboValue.get("CLASS_GROUP_NAME"), "0", "1");
+			if (expectedTurboValue.size() > 1) {
+				rightClick("clickOnCheckbox");
+				editClassificationListForTurboAndDiesel("diesel", "N");
+				String checkboxNo = getAttributeValue("clickCheckBoxOfDiesel", "Toggle.ToggleState");
+				HashMap<String, String> expectedTurboValueFinal = getClassificationInfomation(
+						actualTurboValue.get("CLASS_GROUP_NAME"), "0", "0");
+				if (expectedTurboValueFinal.size() > 1)
+					assertEquals(checkboxNo, "0");
+				else
+					assertEquals(false, true);
+
+			} else {
+				throw new Exception("Case failed");
+			}
+		} else {
+			throw new Exception("checkbox is not clicked");
+		}
 	}
 }
